requirements.txt
# Core Django dependencies
Django==5.2.1
djangorestframework==3.16.0

# Authentication and Authorization
djoser==2.3.1
djangorestframework_simplejwt==5.5.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1

# Image processing and media files
Pillow==10.4.0

# Filtering and search
django-filter==24.3

# CORS support for frontend integration
django-cors-headers==4.7.0

# Database and ORM
psycopg2-binary==2.9.9

# Additional dependencies
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0

# Development and testing (optional)
# pytest==8.3.2
# pytest-django==4.9.0
# factory-boy==3.3.1
